forensics and virus scanning/requirements.txt
PyQt5>=5.15
Pillow>=10.0
PyPDF2>=3.0
python-magic-bin; platform_system == "Windows"
python-magic; platform_system != "Windows"
exifread>=3.0
pefile>=2023.2.7; platform_system == "Windows"
lief>=0.14.1; platform_system != "Windows"
reportlab>=4.0
python-dotenv>=1.0
pywin32>=306; platform_system == "Windows"